refactor: 프로젝트 코드 전체 개선
