+++ src/main/java/roomescape/exception/AccessDeniedException.java
@@ -5,8 +5,4 @@ public class AccessDeniedException extends RuntimeException {
     public AccessDeniedException(final String message) {
         super(message);
     }
-
-    public AccessDeniedException(final String message, final Throwable cause) {
-        super(message, cause);
-    }
 }
