+++ src/main/java/roomescape/exception/UnauthorizedException.java
@@ -5,8 +5,4 @@ public class UnauthorizedException extends RuntimeException {
     public UnauthorizedException(final String message) {
         super(message);
     }
-
-    public UnauthorizedException(final String message, final Throwable cause) {
-        super(message, cause);
-    }
 }
