+++ src/main/java/roomescape/reservation/controller/AdminReservationController.java
@@ -8,13 +8,13 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.reservation.controller.response.ReservationResponse;
-import roomescape.reservation.controller.response.ReservationTimeResponse;
-import roomescape.reservation.controller.request.SaveReservationRequest;
-import roomescape.reservation.controller.request.SaveReservationTimeRequest;
-import roomescape.reservation.controller.request.SaveThemeRequest;
-import roomescape.reservation.controller.request.SearchReservationsRequest;
-import roomescape.reservation.controller.response.ThemeResponse;
+import roomescape.reservation.dto.ReservationResponse;
+import roomescape.reservation.dto.ReservationTimeResponse;
+import roomescape.reservation.dto.SaveReservationRequest;
+import roomescape.reservation.dto.SaveReservationTimeRequest;
+import roomescape.reservation.dto.SaveThemeRequest;
+import roomescape.reservation.dto.SearchReservationsRequest;
+import roomescape.reservation.dto.ThemeResponse;
 import roomescape.reservation.dto.ReservationDto;
 import roomescape.reservation.dto.ReservationTimeDto;
 import roomescape.reservation.dto.ThemeDto;
