+++ src/main/java/roomescape/reservation/controller/ReservationController.java
@@ -6,9 +6,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.auth.principal.AuthenticatedMember;
-import roomescape.reservation.controller.response.MyReservationResponse;
-import roomescape.reservation.controller.response.ReservationResponse;
-import roomescape.reservation.controller.request.SaveReservationRequest;
+import roomescape.reservation.dto.MyReservationResponse;
+import roomescape.reservation.dto.ReservationResponse;
+import roomescape.reservation.dto.SaveReservationRequest;
 import roomescape.reservation.dto.ReservationDto;
 import roomescape.reservation.service.ReservationService;
 import roomescape.resolver.Authenticated;
