+++ src/main/java/roomescape/reservation/controller/ThemeController.java
@@ -3,8 +3,8 @@ package roomescape.reservation.controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.reservation.controller.response.AvailableReservationTimeResponse;
-import roomescape.reservation.controller.response.ThemeResponse;
+import roomescape.reservation.dto.AvailableReservationTimeResponse;
+import roomescape.reservation.dto.ThemeResponse;
 import roomescape.reservation.service.ThemeService;
 
 import java.time.LocalDate;
