+++ src/main/java/roomescape/reservation/dto/SaveReservationRequest.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.controller.request;
+package roomescape.reservation.dto;
 
 import roomescape.member.model.Member;
 import roomescape.reservation.model.Reservation;
