+++ src/main/java/roomescape/reservation/dto/SaveReservationTimeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.controller.request;
+package roomescape.reservation.dto;
 
 import roomescape.reservation.model.ReservationTime;
