+++ src/test/java/roomescape/reservation/service/ReservationServiceTest.java
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
-import roomescape.reservation.controller.request.SaveReservationRequest;
+import roomescape.reservation.dto.SaveReservationRequest;
 import roomescape.reservation.dto.ReservationDto;
 
 import java.time.LocalDate;
