+++ src/test/java/roomescape/reservation/service/ReservationTimeServiceTest.java
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
-import roomescape.reservation.controller.request.SaveReservationTimeRequest;
+import roomescape.reservation.dto.SaveReservationTimeRequest;
 import roomescape.reservation.dto.ReservationTimeDto;
 
 import java.time.LocalTime;
